move hotspot editor into own file
